+++ cmd/varlogadm/main.go
@@ -6,15 +6,15 @@ import (
 
 	_ "go.uber.org/automaxprocs"
 
-	"github.com/kakao/varlog/cmd/vms/app"
-	"github.com/kakao/varlog/internal/vms"
+	"github.com/kakao/varlog/cmd/varlogadm/app"
+	"github.com/kakao/varlog/internal/varlogadm"
 )
 
 func main() {
-	options := vms.DefaultOptions()
+	options := varlogadm.DefaultOptions()
 	app := app.InitCLI(&options)
 	if err := app.Run(os.Args); err != nil {
-		log.Printf("vms: %v", err)
+		log.Printf("varlogadm: %v", err)
 		os.Exit(1)
 	}
 }